feat: Initialize uInspESP32 project structure with essential files
This commit sets up the initial structure for the uInspESP32 project, including the addition of a .gitignore file to exclude build artifacts, a README.md for project documentation, and configuration files for PlatformIO. It also introduces core source files such as RingBuf.hpp, SimpPacketParse.hpp, and SyncTask.cpp, laying the groundwork for the micro inspection system controller's functionality. These changes establish a foundation for further development and modularization of the firmware.